// I_Interactable Interface Implementation + Key and Lock System Upgrade
I created an interface that handles the events associated with the Interact Key. When the player presses the Interact key, the interface will be called on the colliding actor. If that actor contains a method for Interact, that method will execute.

I upgraded the Lock and Key System to allow the player to hold multiple keys at once.